requirements.txt
flask==3.0.3
requests==2.32.3
mysql-connector-python==8.4.0
